Add workflows to test with mainnet protocol config
## Description 

It is common to have a feature enabled in simtests and optionally
testnet, but not mainnet. A feature can be in this state for months. But
mainnet protocol config is only tested in a few upgrade related tests.
It is important to make sure changes do not accidentally break the
`mainnet` protocol while passing tests with the latest protocol config
via features enabled in dev environments.

## Test plan 

CI

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: