[trace-view] Fixed event emitting problem
## Description 

This PR fixes a problem related to re-using runtime-level functions to
implement certain actions. For example, `stepOut` function was
implemented by repeatedly calling `next` function.

The problem was that these these re-used functions were also called from
the adapter and where expected to send and even to the client informing
it of the need to suspend execution in the debugger. Clearly this event
should only be send once per completed action but because the functions
implementing actions were re-used (in loops) an excessive number of such
events would be send, slowing down the debugger or even stalling it
permanently in extreme cases.

The solution is to refactor action implementation to private internal
functions that do not send events to the client, re-use these internal
functions in the implementation, and create top-level public
event-sending functions to be used by the adapter.

This PR also includes a very small improvement to stack request handling
that's unrelated to the main goal of this PR, and was included to avoid
fragmenting the review process too much as the content of this PR
remains rather small even with the two fixes included

## Test plan 

All existing tests must pass. Also tested manually to make sure that
there is not excessive communication of events to the client