[execution] Pass early_execution_error to execution layer
## Description 

Previously we determine if a transaction should fail execution
immediately for various reasons within the execution layer. This
requires us passing extra information through a few layers. And it makes
it hard to add more types of those failures, for instance, insufficient
balance withdraws.
This PR changes this to do the decision making outside, and put all code
in the same file so it's clear.

## Test plan 

CI
Should be a non-functional change.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: