Observe and commit to "indirect" state.
Indirect state is any state we inspect and make a decision on while
processing a commit, which could diverge unintentionally due to code
changes.

The contents of a commit are already committed to, but any non-trivial
derivation of that information (for instance, reading from persistent
state that was constructed while processing previous commits) is not,
prior to this change.

This means that divergences in state can hide until they cross some
threshold, which means both that we are less likely to catch the issue
in tests, and also that a fork may happen long after that state has
diverged, rather than immediately, which makes the fork much harder to
repair.