fix(zklogin): skip invalid jwk not error
## Description 

we used to error completely if a provider has a ES256 key in addition to
a RSA key. it should instead skip it and parse the valid RSA key. added
to protocol config flag:
https://github.com/GeunhwaJeong/fastcrypto/pull/858
## Test plan 

How did you test the new or updated feature?

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: