display(v2): stricter handling of malformed UTF-8
## Description

Make the lexer and parser more resilient to non-UTF-8 characters, by
only accepting ASCII whitespace, and improving our handling of
unexpected bytes to be sensitive to the possibility of bad UTF-8.

This requires improvements to both the unexpected token detection, and
also unexpected token printing (both for test and for error generation).

## Test plan

```
$ cargo nextest run -p haneul-display
```