[move-ide] Fixed cached and new program components merge
## Description 

This PR fixes merge between new and cached program components in the
face of incremental compilation. The new algorithm is really simple. As
input, we have a list of modified files, and a mapping from hashes to
file paths for both the new version of the (compiled) program and the
cached version of the (compiled) program. In order to determine if the
module has been modified or not (and thus whether we need a new version
or the cached version), we get the path for the module based on its hash
from the respective map and check if that path is in the modified files
set

## Test plan 

Tested manually that the whole thing works correctly now