Fix flaky test by adding retry logic to publish_basics_package
## Description 

In simtests it is possible that ValidatorTxFinalizer (5s delay) is able
to finalize a publish of a benchmark package before a client retry is
successful in a very small window before the epoch terminates. This
package is required for workloads to run. The publish tx originally
failed due to ValidatorHaltedAtEpochEnd as the timing of when the epoch
(5s epoch) changed and this benchmark init code was run coincided. The
benchmark init code does not tolerate any failures with package
publishing and it causes the whole test to fail which should be fixed.
The reason it only has come up now is that TD/QD being mixed allows for
benchmark setup to run at a different timings than when it was 100% QD
which opened up this window for the failure to happen.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: