indexer-alt: simplify reader error-handling
## Description

Simplify the shared error type used by `haneul-indexer-alt-reader` to a
simple wrapper over `anyhow` that allows for the error to be shared
across multiple requests that came to a data-loader.

Previously this error was semi-structured with different variants for
postgres and bigtable, and for creation vs querying. However, in all
cases these errors made their way into an `anyhow::Error`, so it was
simplified as such.

This is in preparation for introducing the consistent-store based
reader, which will need a more structured error type (but which
nevertheless probably should not be shared among all reader types).

## Test plan

```
$ cargo nextest run          \
  -p haneul-indexer-alt-reader  \
  -p haneul-indexer-alt-jsonrpc \
  -p haneul-indexer-alt-graphql \
  -p haneul-indexer-alt-e2e-tests
```