graphql-alt: consistent store watermark
## Description

Bound reads from GraphQL by the watermarks from the consistent store, if
the consistent reader has been configured.

Note that unlike the postgres and bigtable watermark reads, which
happens concurrently, the consistent store watermark reads happens at
the end. This is because it accepts the watermark from the other
services as input. We do this to support reading from a consistent store
whose stride is greater than 1.

## Test plan

Existing E2E tests:

```
$ cargo nextest run -p haneul-indexer-alt-e2e-tests
```