[move-compiler] A leaner version of pre-compiled libs
## Description 

This PR replaces `FullyCompiledProgram` used for testing and in to speed
up compilation in `move-analyzer` with a much leaner per-module
`PreCompiledModuleInfo` structure.

The effect on `move-analyzer`'s memory consumption is quite dramatic -
for a single package with default dependencies, memory footprint shrinks
to ~350M vs. 1.4GB before this optimization was implemented.

The changes to `move-analyzer` are more involved than simply swapping
`FullyCompiledProgram` with `PreCompiledModuleInfo` for the following
reason. Prior to this PR we could use pre-compiled dependencies during
first compilation as it contained full parsing and typing ASTs that
could be directly plugged into the analysis even though the pre-analyzed
dependencies were not available during this first compilation. Now we no
longer have full ASTs available so we do the first compilation without
pre-compiled dependencies (so that full ASTs for both user program and
dependencies can be produced), and use both pre-compiled dependencies
(for faster compilation) and pre-analyzed dependencies (to avoid
re-analyzing them) during subsequent compilations. This results in the
first compilation to be slower but I'd say it's an acceptable trade-off

## Test plan 

All CI tests must pass. Tested `move-analyzer` modifications manually