enforce 0600 flags on haneul.keystore
## Description 

Ensure that (like SSH) we reduce the file permissions on haneul.keystore.

Note that this will automatically alter the flags of the haneul.keystore
file upon load or save.

This change does not apply to Windows.

## Test plan 

CI Pipeline and manual testing.

---

## Release notes

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes):
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [x] CLI: (linux, macos) Ensure that haneul.keystore files are flagged
with 0600 flags. This is a defense-in-depth measure to defend Haneul
wallets running on shared multi-user hosts.
- [ ] Rust SDK: