Enable Accumulator Version Discovery for Balance   Withdrawals
## Description 

This PR adds the necessary infrastructure to discover and assign
accumulator versions for balance withdraw transactions through
consensus.
We do this as part of the version assignment process, and return the
accumulator version in the assignment result if it's a withdraw
transaction.

## Test plan 

added unit tests

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: