graphql-alt: propagate package resolver through scope
## Description

Rather than expose the package resolver through the `Context` (which
implies using the same resolver for the whole GraphQL request), expose
it through `Scope`. This allows us to override it within nested queries,
which allows us to support package resolution for requests that involve
resolving types from a package that was published within that
transaction.

## Test plan

Existing tests:

```
cargo nextest run              \
  -p haneul-indexer-alt-e2e-tests \
  -p haneul-indexer-alt-graphql   \
  -p haneul-indexer-alt-jsonrpc
```