Small refactors in commit handler to prepare for checkpoint splitting changes
In order to split checkpoints in consensus handler, we need to assign
checkpoint heights dynamically. This in turn means that we need to
create `Schedulable::AccumulatorSettlement` close to where we create
pending checkpoints so that. This is just a small refactor to prepare
for that, to keep subsequent PRs smaller/less cluttered.