Remove revert_state_update - clear_state_end_of_epoch reverts all state
This has been a long-standing TODO since writeback cache landed. This
method is simpler and clearly correct: Anything still in the dirty set
at the end of the epoch must not have been in a checkpoint. The only
tricky part is making sure to correctly invalidate the caches for
anything in the dirty set.