graphql-alt: sort schema fields
## Description 

This sorts the fields in the generated GraphQL schema lexicographically
instead of depending on the ordering of the fields/methods in rust.

Based on this comment
https://github.com/GeunhwaJeong/haneul/pull/23194#discussion_r2285311478.

## Test plan 

Only generated schema is affected.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: