graphql-alt: name service APIs
## Description

Add APIs to map a HaneulNS name to an `Address`, and return the default
HaneulNS name for an address, taking into account expiration.

The implementation is based on the equivalent API in `jsonrpc-alt`, but
accounting for GraphQL's methods for fetching objects, in a consistent
way.

## Test plan

New E2E tests -- these need to be implemented in Rust, so that a custom
name service config can be set:

```
$ cargo nextest run            \
  -p haneul-indexer-alt-e2e-tests \
  --test graphql_name_service_tests
```

The tests are adapted from the equivalent tests for `jsonrpc-alt`.