Check conservation of settlement transactions correctly.
First, we need to make HANEUL contained in accumulator objects visible
so that it can be accounted for in the conservation check that happens
at reconfiguration, when all objects are scanned.

Once we do that, it then appears that settlement transactions are
minting HANEUL. Settlement transactions gather HANEUL deposited or
withdrawn from the effects of other transactions, and write it to
accumulator objects. But execution cannot see where this HANEUL came
from, so it appears that the transaction is minting HANEUL, and the
conservation checks fail.

To address this, we pass the total input and output HANEUL to the
settlement prologue entry point.  These numbers are then recorded
in the object runtime and eventually propagated to the TemporaryStore
so that they can be accounted for in conservation checks.