[TransactionDriver] set timeout from benchmark workload
## Description 

Use timeout in benchmark workload, so drive_transaction() will not get
stuck when it is unable to finalize a transaction but the error is
retriable (for example, 30% stake of lock conflicts, 4% stake of
unavailable validators / timeouts).

Refactor local validation error.

## Test plan 

CI

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: