graphql-alt: Object.dynamicField et al
## Description

Add dynamic fields to `Object`. Previously these functions existed on
`Address` and on `MoveObject` but not on `Object`, but I spotted based
on the example queries that we did used to have it there.

The reason it's not part of the `IAddress` interface is that this would
require us to implement it on `MovePackage` where it would be entirely
redundant.

## Test plan

Updated E2E tests:

```
$ cargo nextest run -p haneul-indexer-alt-e2e-tests -- graphql/objects/dynamic_field
```

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: