Rename `CallArg::BalanceWithdraw` and related names to match `funds_accumulator`
## Description 

- `CallArg::BalanceWithdraw` items have been renamed to
`FundsWithdrawal` or similarly titled names. The type of this `CallArg`
in the PTB will always be `haneul::funds_accumulator::Withdrawal` and the
`CallArg` variant was renamed with this in mind.
  - See  for details on names of members `haneul::funds_accumulator` 
- Some other small naming consistency changes

## Test plan 

- clippy

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: