[tidehunter] relocation for perpetual db
## Description 

Initial version of Tidehunter pruning for perpetual DB
The pruning works based on relocation filters (similar to RocksDB's
compaction filters).
Every epoch, the pruner updates the target epoch watermark (equal to
current epoch – number of epochs to retain in config) and triggers
Tidehunter's relocation.
The corresponding relocation filters check the data present in a
key–value pair and, if it corresponds to an eligible epoch, remove it.

The key difference from RocksDB's compaction is that the relocation
process here is sequential, and all records are guaranteed to be
processed in the order they were added to the WAL. This makes it
possible to prune entries that don’t necessarily carry epoch
information. For instance, transactions and events can be pruned
unconditionally, because they are guarded by the corresponding effects
entry. If the relocator sees any entry from an epoch that is not
eligible for pruning, it pauses the whole relocation process

Extra notes:
* current version does not have a smoother implemented

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: