[pkg-alt] Fix git sparse checkout
## Description 

This PR fixes the sparse checkout command. The issue was that
`sparse-checkout add .` does not do anything and in particularly, does
not add any folders to be checked out. The fix was to manually edit the
`.git/info/sparse-checkout` file to have a `*` when we need to check out
the whole repo because no path is given, or the specific path provided
in the toml.

## Test plan 

Add new tests
`cargo nextest run -p move-package-alt`.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: