[3/n][haneul-adapter] Update (but dont enable) object runtime to use abstract memory size
## Description 

Update the last remaining calls to `legacy_size` to use
`abstract_memory_size`.

This means that we will no longer have any calls to `legacy_size` in the
next execution version (🥳), and be able to remove it at that time. We
should be able to actually convert all the `legacy_size` calls over to
`abstract_memory_size` calls today I believe without gating, but decided
to keep the old function around + protocol gate just to be on the safe
side....

The changes are protocol gated and the protocol config is not enabled.
That will be done at a later date :)

## Test plan 

CI -- nothing should change.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: