[MFP] Ability to receive ping transactions on the server
## Description 

This is the first of a set of 4 PRs to implement the end to end latency
checks for the transaction driver. The list of all the PRs is here:

* https://github.com/GeunhwaJeong/haneul/pull/23543
* https://github.com/GeunhwaJeong/haneul/pull/23544
* https://github.com/GeunhwaJeong/haneul/pull/23545

This PR though is introducing the changes need on the server side to
support sending a ping transaction in order to properly measure the end
to end latency for testing separately the `Fast Path` and `Consensus`.
We are basically simulating the ping transaction (never gets included in
the block) and also re-using the `ConsensusTxStatusCache` . Instead of
tracking the block status (certified/finalised) , we do that via a
simulated transaction by reserving a position (index) in the block.

For this the `submit_transaction` and `wait_for_effects` rpc endpoints
have been reused. I've tried to avoid as much as possible the implicit
assumptions to prioritise safety.

## Test plan 

CI

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:

---------