[MFP] Replace the moving decay window with moving window
## Description 

Replacing the moving decay window to moving window. From experimentation
when trying to measure the end to end (finalisation) latency the using
decay window is hard to tune and also susceptible to jitters.

The PR is also measuring the end to end latency (from tx submission up
to gathering the certified effects) and is tuning the configuration so
we take into account mostly the `Finalization` latency

## Test plan 

CI