[move][move-stackless-bytecode] Refactor stackless bytecode repository to simplify logic
## Description 

This cleans up a lot of the testing and structure of the stackless
bytecode, hopefully toward a nicer future.

## Test plan 

Test revisions.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: