[MFP] enhance metrics with additional labels. Refactor tx retry back off
## Description 

This PR:

* Enhances some metrics with additional labels
* Refactors the back off strategy when retrying the tx submission

Running for `10` times the back off with the current settings yields
results like:

```
58.618524ms
1.717898302s
626.285083ms
8.890868961s
6.853596566s
1.27459605s
8.60306108s
2.898520074s
1.538756239s
2.389480396s
```

Although we are using an exponential back off, the current setup and
jitter seems to mess up a bit the back off latency to be used. It's not
very scalar in the sense that we should expect using a bigger and bigger
step every time.

With the proposed changes we'll have something like:

```
136.153155ms
172.376361ms
315.614501ms
460.543044ms
866.642052ms
1.196423s
1.584154394s
3.647710108s
4.406361196s
8.318059196s
```

## Test plan 

How did you test the new or updated feature?

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: