[graphql-alt] Support newly created objects for PackageResolver
## Description 

With `Query::simulateTransaction` and `Mutation::execute_transaction`,
we implemented the `execution_objects` to Scope to let us load the
object from just-executed transactions instead of querying the DB. This
PR applies that `execution_objects` to `PackageResolver`, which could be
helpful in case a resolution request comes in for a type in a package
that was just published.

## Test plan 

How did you test the new or updated feature?

```
 cargo nextest run -p haneul-indexer-alt-e2e-tests
```

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: