[MFP] Refactor the use of scoring - use latencies instead
## Description 

The current scoring mechanism is working well but it is less intuitive
and it currently contains as well unnecessary calculations. An
alternative is to use instead the calculated average latencies (end to
end) and expose those as the "scores". The reliability metric will now
be used to penalise the latency with the max latency been chosen the
`10s` - anything above that doesn't have much meaning anyways.

I've also steered away from re-using the "score" term so it becomes
clear end to end what we are measuring here.

## Test plan 

CI/PT

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:

---------