[move][move-model] Allow missing dependencies via configuration
## Description 

This adds a `ModelBuilderConfig` and a flag to allow missing
dependencies when generating a move model.

## Test plan 

Testing was done downstream in the decompiler.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: