[MFP] remove the failure exclusion logic
## Description 

Given the latest refactor that penalises the validator's end to end
adjusted latency, the logic to exclude a validator with an X number of
consecutive failures seems a bit redundant. A very unreliable validator
which responds with errors will get ranked with a poor score and won't
be selected for tx submission anyways. Having this additional logic here
adds more complexity which practically doesn't provide much benefit.

The ping transactions will play the role of recovering their
reliability. Maybe the current exclusion logic could protect us from
oscillations, as it gives some time before we actually prefer the
validator again to submit transactions to, but I believe this will be
covered by the fact that sampling will not be that frequent anymore and
pings (which kick in every 15seconds) will be the ones that will give
that opportunity.

## Test plan 

CI/PT

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: