indexer-alt: expose remote ingestion client
## Description

Make the remote ingestion client available for users outside the crate.
This is in preparation for supporting restore from formal snapshot which
requires access to a remote store to map epochs to their last
checkpoint.

Among the ingestion clients, only the remote client has this
information, so it makes sense to expose that rather than expand the
ingestion client interface to include this information.

## Test plan

Existing tests.