bugfix(consistency): Correctly index Effects V1 input objects
## Description

In Effects V1, the input state of modified objects doesn't include the
digest. The consistent store expected the input state to have both the
version and the digest, so it wasn't correctly deleting the previous
version of the modified object.

## Test plan

New regression test:

```
$ cargo nextest run            \
  -p haneul-indexer-alt-e2e-tests \
  --test consistent_store_list_owned_objects_tests
```

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [x] GraphQL: Fix a bug where live object set queries would include historical data.
- [ ] CLI: 
- [ ] Rust SDK: