[replay] Added support for replacing cached packages
## Description 

This PR adds support for replacing packages in the cache, allowing to
replay over modified (with respect to code being executed) local state.

It adds three commands:
- `rebuild-package` - takes package ID and package source path
(optionally also node if different from mainnet) as parameters and
locally constructs a package BCS binary from provided sources; if output
path is specified, writes binary to this path, otherwise replaces
package in the file cache
- `extract-package` - takes package ID and output path (optionally also
node if different from mannet) as parameters and extracts a given
package from file cache to save it at a given path
- `overwrite-package` - takes a path to a package BCS binary and package
ID (optionally also node if different from mainnet) as parameters and
writes package BCS binary into the file cache

This PR also reworks how forked artifacts are handled as it was somewhat
confusing before. This is how it works now:
1. on first transaction replay
- if effects from on chain are the same as the replayed ones, write
transaction effects artifact
- otherwise, write replayed effects as forked effects artifact and
on-chain effects as "regular" effects artifacts (and don't error on
fork)

This way, even at the first execution, we will see either one set of
effects (if no fork) or a diff (if forked)

2. on subsequent transaction replay
- if effects from on chain are the same as the replayed ones, write
transaction effects artifact, but also remove forked effects artifact so
that no (spurious) diff is reported (after all local replay and on-chain
effects are the same)
- otherwise, same as during first execution, write replayed effects as
forked effects artifact and on-chain effects as "regular" effects
artifacts

This way, if we keep changing replay (due to providing different package
versions), we will keep seeing effects diff for as long as replayed
effects are different from on-chain effects. If we happen to hit the
same effects locally as they are on chain, we will go back to "regular"
output and no diff

## Test plan 

Tested manually