[move-ide] Fixed incorrectly inserted macro-related uses
## Description 

This PR fixes a problem with uses coming from inlined macro calls being
spuriously added in random location causing go-to-def, on-hover, etc. to
go haywire.

The reason was that the original algorithm assumed that all symbols
processed in a given module are located in this module (and maintained a
map of use-defs on a per-module basis). This is of course not true in
presence of macros, and the algorithm was changed (and simplified
somewhat in the process) to keep use-defs on a per-file basis throughout
the analysis.

## Test plan 

Added an additional test. All tests must pass