Update submask filter for balance_change in simulate transaction
## Description 

Currently, the balance change is not returned in gRPC
simulateTransaction when we defined the submask
`transaction.balance_changes`.

The culprit is the filtering logic is not using `transaction` subtree,
and it happened for both v2 and v2beta.

## Test plan 

How did you test the new or updated feature?

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: