[indexer-alt-framework] use min next seq checkpoint to compute initial ingest hi
## Description 

Currently the ingest_hi is initialized to None, which could cause
checkpoints to be sent to sequential pipelines without any regulation
before any commit has been made. This PR adds an initial value to make
sure there's regulation from the very beginning.
Additionally to align with the `next_checkpoint` change, I've made the
value exclusive and changed the name in various places to `commit_hi` to
make it more clear that this is the highest commit watermark the
pipelines are going to reach next.

## Test plan 

Added tests.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: